+++ lib/budget/savings/saving.ex
@@ -3,9 +3,10 @@ defmodule Budget.Savings.Saving do
   import Ecto.Changeset
   alias Budget.Savings.Saving
 
+
   schema "savings" do
-    field(:amount, :decimal)
-    field(:title, :string)
+    field :amount, :float
+    field :title, :string
 
     timestamps()
   end
